--- src/pages/AuthorizationPage/SignIn/index.ts
@@ -1,1 +1,0 @@
-export { default } from './signIn';
