--- src/pages/AuthorizationPage/SignUp/index.ts
@@ -1,1 +1,0 @@
-export { default } from './signUp';
