--- src/pages/AuthorizationPage/index.ts
@@ -1,1 +1,0 @@
-export * from './types';
