+++ src/pages/SignUpPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './signUpPage';
